Split up sidebar into multiple components
