+++ src/server/DasCleverle.DcsExport.LiveMap/ClientApp/src/components/Map/Map.js
@@ -1,10 +1,12 @@
-import React, { useState } from 'react';
+import React, { useCallback, useState } from 'react';
 import { useStore } from 'react-redux';
 
 import { Mapbox } from './Mapbox';
-import { Sidebar } from './Sidebar';
+import { Sidebar } from './Sidebar/Sidebar';
+import { MissionSidebarCard } from './Sidebar/MissionSidebarCard';
+import { AirbaseSidebarCard } from './Sidebar/AirbaseSidebarCard';
 
-const ObjectPopup = (selectedObjects) => {
+const ObjectPopup = ({ selectedObjects }) => {
   return (
     <div className="mt-2">
       {selectedObjects.map((o) => (
@@ -21,37 +23,43 @@ export function Map() {
   const store = useStore();
   const [selectedAirbase, setAirbase] = useState(null);
 
-  const handleMapClick = (map, layer, features) => {
-    const {
-      liveState: { airbases, objects },
-    } = store.getState();
+  const handleMapClick = useCallback(
+    (map, layer, features) => {
+      const {
+        liveState: { airbases, objects },
+      } = store.getState();
 
-    switch (layer) {
-      case 'airbases':
-        if (features.length === 0) {
-          setAirbase(null);
-        } else {
-          setAirbase(airbases[features[0].properties.id]);
-        }
-        break;
+      switch (layer) {
+        case 'airbases':
+          if (features.length === 0) {
+            setAirbase(null);
+          } else {
+            setAirbase(airbases[features[0].properties.id]);
+          }
+          break;
 
-      case 'objects':
-        const selectedObjects = features.map((f) => objects[f.properties.id]);
-        if (selectedObjects.length === 0) {
-          return null;
-        }
+        case 'objects':
+          const selectedObjects = features.map((f) => objects[f.properties.id]);
+          if (selectedObjects.length === 0) {
+            return null;
+          }
 
-        return ObjectPopup(selectedObjects);
+          return <ObjectPopup selectedObjects={selectedObjects} />;
 
-      default:
-        break;
-    }
-  };
+        default:
+          break;
+      }
+    },
+    [store]
+  );
 
   return (
     <>
       <Mapbox onClick={handleMapClick} />
-      <Sidebar selectedAirbase={selectedAirbase} />
+      <Sidebar>
+        <MissionSidebarCard />
+        <AirbaseSidebarCard airbase={selectedAirbase} />
+      </Sidebar>
     </>
   );
 }
