+++ src/server/DasCleverle.DcsExport.LiveMap/ClientApp/src/components/Map/Sidebar/Sidebar.css
