+++ src/server/DasCleverle.DcsExport.LiveMap/ClientApp/src/components/Map/Sidebar/Sidebar.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import './Sidebar.css';
+
+export function Sidebar({ children }) {
+  return <div className="sidebar">{children}</div>;
+}
